feat(webpack): add NxWebpackPlugin that works with normal Webpack configuration
