+++ e2e/webpack/src/webpack.test.ts
@@ -1,6 +1,7 @@
 import {
   cleanupProject,
   newProject,
+  packageInstall,
   rmDist,
   runCLI,
   runCommand,
@@ -11,8 +12,8 @@ import {
 import { join } from 'path';
 
 describe('Webpack Plugin', () => {
-  beforeEach(() => newProject());
-  afterEach(() => cleanupProject());
+  beforeAll(() => newProject());
+  afterAll(() => cleanupProject());
 
   it('should be able to setup project to build node programs with webpack and different compilers', async () => {
     const myPkg = uniq('my-pkg');
@@ -86,7 +87,7 @@ module.exports = composePlugins(withNx(), (config) => {
 
     updateFile(
       `libs/${myPkg}/.babelrc`,
-      `{ "presets": ["@nx/js/babel", "./custom-preset"] } `
+      `{ 'presets': ['@nx/js/babel', './custom-preset'] } `
     );
     updateFile(
       `libs/${myPkg}/custom-preset.js`,
@@ -105,5 +106,33 @@ module.exports = composePlugins(withNx(), (config) => {
       },
     });
     expect(output).toContain('Babel env is babelEnv');
+  }, 500_000);
+
+  it('should be able to build with NxWebpackPlugin and a standard webpack config file', () => {
+    const appName = uniq('app');
+    runCLI(`generate @nx/web:app ${appName} --bundler webpack`);
+    updateFile(`apps/${appName}/src/main.ts`, `console.log('Hello');\n`);
+
+    updateFile(
+      `apps/${appName}/webpack.config.js`,
+      `
+      const path  = require('path');
+      const { NxWebpackPlugin } = require('@nx/webpack');
+
+      module.exports = {
+        target: 'node',
+        output: {
+          path: path.join(__dirname, '../../dist/${appName}')
+        },
+        plugins: [
+          new NxWebpackPlugin()
+        ]
+      };`
+    );
+
+    runCLI(`build ${appName} --outputHashing none`);
+
+    let output = runCommand(`node dist/${appName}/main.js`);
+    expect(output).toMatch(/Hello/);
   }, 500_000);
 });
