+++ packages/webpack/src/plugins/nx-webpack-plugin/nx-webpack-plugin-options.ts
@@ -1,3 +1,4 @@
+import { ProjectGraph } from '@nx/devkit';
 import { AssetGlob } from '@nx/js/src/utils/assets/assets';
 
 export interface AssetGlobPattern {
@@ -36,58 +37,64 @@ export interface OptimizationOptions {
   styles: boolean;
 }
 
-export interface WebpackExecutorOptions {
+export interface NxWebpackPluginOptions {
+  // Required options
+  main: string;
+  outputPath: string;
+  tsConfig: string;
+
+  // Optional options
   additionalEntryPoints?: AdditionalEntryPoint[];
   assets?: Array<AssetGlob | string>;
-  buildLibsFromSource?: boolean;
+  babelConfig?: string;
+  babelUpwardRootMode?: boolean;
+  baseHref?: string;
   commonChunk?: boolean;
   compiler?: 'babel' | 'swc' | 'tsc';
+  crossOrigin?: 'none' | 'anonymous' | 'use-credentials';
   deleteOutputPath?: boolean;
+  deployUrl?: string;
   externalDependencies?: 'all' | 'none' | string[];
+  extractCss?: boolean;
   extractLicenses?: boolean;
   fileReplacements?: FileReplacement[];
+  generateIndexHtml?: boolean;
   generatePackageJson?: boolean;
-  isolatedConfig?: boolean;
-  main: string;
+  index?: string;
   memoryLimit?: number;
   namedChunks?: boolean;
   optimization?: boolean | OptimizationOptions;
   outputFileName?: string;
   outputHashing?: any;
-  outputPath: string;
   poll?: number;
   polyfills?: string;
+  postcssConfig?: string;
   progress?: boolean;
   runtimeChunk?: boolean;
-  sourceMap?: boolean | 'hidden';
-  statsJson?: boolean;
-  target?: 'node' | 'web' | 'webworker';
-  transformers?: TransformerEntry[];
-  tsConfig: string;
-  vendorChunk?: boolean;
-  verbose?: boolean;
-  watch?: boolean;
-  webpackConfig?: string;
-  babelConfig?: string;
-  babelUpwardRootMode?: boolean;
-  baseHref?: string;
-  crossOrigin?: 'none' | 'anonymous' | 'use-credentials';
-  deployUrl?: string;
-  extractCss?: boolean;
-  generateIndexHtml?: boolean;
-  index?: string;
-  postcssConfig?: string;
   scripts?: Array<ExtraEntryPointClass | string>;
+  skipTypeChecking?: boolean;
+  sourceMap?: boolean | 'hidden';
+  ssr?: boolean;
+  statsJson?: boolean;
   stylePreprocessorOptions?: any;
   styles?: Array<ExtraEntryPointClass | string>;
   subresourceIntegrity?: boolean;
+  target?: string | string[];
+  transformers?: TransformerEntry[];
+  vendorChunk?: boolean;
+  verbose?: boolean;
+  watch?: boolean;
 }
 
-export interface NormalizedWebpackExecutorOptions
-  extends WebpackExecutorOptions {
-  outputFileName: string;
-  assets: AssetGlobPattern[];
+export interface NormalizedNxWebpackPluginOptions
+  extends NxWebpackPluginOptions {
+  projectName: string;
   root: string;
   projectRoot: string;
   sourceRoot: string;
+  configurationName: string;
+  targetName: string;
+  projectGraph: ProjectGraph;
+  outputFileName: string;
+  assets: AssetGlobPattern[];
 }
